deprecate architectures folder in test
