+++ test/models/test_late_fusion.py
